Re-organize the guide about optimized and generic builds
When downloading Zen, most users will wonder what's the difference between the optimized and generic
build.

The previous version of this topic had that information at the end, which causes users to miss
the information.

I restructured this topic to have that information in a more prominent location. I also simplified
the list as having listed every single supported family is not that helpful (most people will have
issues locating their CPUs in such a long list)